pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-organizer"
version = "1.0.0"
description = "A tool for organizing and deduplicating image and video files"
authors = [{name = "Image Organizer Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
    "piexif>=1.1.3",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
imgtool = "imgtool.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "PIE", "T20", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "PTH", "TID", "TCH", "INT", "ICN", "G", "INP", "PIE", "LOG", "NPY", "AIR", "PERF", "FBT", "BLE", "COM", "C90", "DTZ", "ISC", "ICN", "G004", "SIM", "TID", "TCH", "ARG", "PIE", "LOG", "NPY", "AIR", "PERF", "FBT", "BLE", "COM", "C90", "DTZ", "ISC", "ICN", "G004"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --strict-config" 